third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-shapes/spec-examples/shape-outside-011-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Shape from image - url(), alpha channel, opacity > 0 Uncaught Error: assert_approx_equals: Line 0 is positioned properly expected 226 +/- 2 but got 216
Harness: the test ran to completion.
